task: leaving Run mechanics within the CLI package and simplifying the CLI implementation with it.
